SIPA-#14_02_Create_Service_for_Product_update - change from Integer to Long

+++ src/main/java/com/example/onlineshop/model/Product.java
@@ -1,6 +1,6 @@
 package com.example.onlineshop.model;
 
-public record Product(Integer id, String title, String description, Double price,Long categoryId) {
+public record Product(Long id, String title, String description, Double price,Long categoryId) {
 
     @Override
     public String toString(){
